Fixed memory leak caused by not calling Peeker.Stop
Not calling stop caused the peeker to stay active until OOM happened,
this was caused by the peeker buffering all data in memory in the case
of a reset. Stop causes the wrapper to switch the reader to the normal
connection when the peeker is exhausted.

Added Close method to RingBuffer, this causes the RingBuffer to return
an io.EOF when empty and closed.